# repo/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Philatelic Museum Operations Suite API
  version: 1.0.0
  description: |
    Public HTTP API for museum operations.

    Authentication model:
    - Session cookie `museum_sid` (set on `POST /auth/login`)
    - `X-CSRF-Token` required on mutating requests (`POST|PUT|PATCH|DELETE`) except `POST /auth/login`
    - `X-Step-Up-Token` required on sensitive actions after `POST /auth/step-up`

    Step-up actions:
    - `GRAPH_PUBLISH`
    - `ROUTE_RULE_CHANGE`
    - `JOB_APPROVE`
    - `JOB_APPEAL_DECIDE`
    - `EXPORT_CREATE`
    - `ADMIN_CONFIG_UPDATE`
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - SessionCookie: []
paths:
...
  /graph/drafts:
    post:
      tags: [Graph]
      summary: Create graph draft
      description: Requires `GRAPH_DRAFT_EDIT` (Administrator, Curator).
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Draft created }
  /graph/drafts/{draftId}:
    get:
      tags: [Graph]
      summary: Read draft snapshot and validation report
      description: |
        Requires `GRAPH_READ`. Unpublished drafts are restricted to Administrator/Curator or draft owner.
      parameters:
        - name: draftId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Draft payload }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /graph/drafts/{draftId}/nodes:
    post:
      tags: [Graph]
      summary: Add node to draft
      description: Requires `GRAPH_DRAFT_EDIT`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '201': { description: Node created }
        '400': { $ref: '#/components/responses/ValidationError' }
  /graph/drafts/{draftId}/nodes/{nodeId}:
    patch:
      tags: [Graph]
      summary: Update node in draft
      description: Requires `GRAPH_DRAFT_EDIT`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
        - name: nodeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Node updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Graph]
      summary: Delete node in draft
      description: Requires `GRAPH_DRAFT_EDIT`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
        - name: nodeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Node deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /graph/drafts/{draftId}/edges:
    post:
      tags: [Graph]
      summary: Add edge to draft
      description: Requires `GRAPH_DRAFT_EDIT`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '201': { description: Edge created }
        '400': { $ref: '#/components/responses/ValidationError' }
  /graph/drafts/{draftId}/validate:
    post:
      tags: [Graph]
      summary: Run draft validation
      description: Requires `GRAPH_DRAFT_EDIT`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Validation report
          content:
            application/json:
              example:
                data:
                  status: INVALID
                  issues:
                    - code: CONSTRAINT_ALLOWED_TARGET_TYPES
                      severity: BLOCKING
                      edgeId: e_a1b2c3d4
                      message: Edge target type ARTIST is not allowed for relation CREATED_BY
  /graph/drafts/{draftId}/publish:
    post:
      tags: [Graph]
      summary: Publish graph draft
      description: Requires `GRAPH_PUBLISH` and `X-Step-Up-Token` for action `GRAPH_PUBLISH`.
      security:
        - SessionCookie: []
          CsrfHeader: []
          StepUpHeader: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Published
          content:
            application/json:
              example:
                data:
                  version: 7
                  publishedAt: '2026-03-30T10:15:00.000Z'
        '422':
          $ref: '#/components/responses/GraphValidationBlocked'
...
  /routes:
    post:
      tags: [Venues & Routes]
      summary: Create route definition
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Route created }
...
  /routes/{routeId}:
    patch:
      tags: [Venues & Routes]
      summary: Update route rules
      description: Requires `ROUTE_RULE_CHANGE` and step-up action `ROUTE_RULE_CHANGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
          StepUpHeader: []
      parameters:
        - name: routeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Route updated }
...
  /exports:
    post:
      tags: [Exports]
      summary: Queue export job
      description: Requires `EXPORT_CREATE` and step-up action `EXPORT_CREATE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
          StepUpHeader: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              resource: participants
              format: CSV
              filters: {}
              fields: [name, phone, email, notes]
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              example:
                data:
                  exportJobId: exp_9fd3a8e2
                  status: QUEUED
  /exports/{exportJobId}:
    get:
      tags: [Exports]
      summary: Export job status/details
      description: |
        Requires `EXPORT_READ` and object-level authorization.
        Only requester or Administrator can access the job.
      parameters:
        - name: exportJobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Export job details }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: museum_sid
      description: Session cookie set by `POST /auth/login`.
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on mutating requests except `POST /auth/login`.
    StepUpHeader:
      type: apiKey
      in: header
      name: X-Step-Up-Token
      description: One-time action-bound step-up token for sensitive operations.
  parameters:
...
    DraftId:
      name: draftId
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              oneOf:
                - type: array
                  items: { type: object }
                - type: object
            requestId: { type: string }
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: Insufficient role permission or object-level access denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    GraphValidationBlocked:
      description: Blocking graph validation issues remain
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }